Update MagneticCave.java
Magnetic Cave Game  is played on an 8x8 chessboard, it is played by two players. The game consists of three modes, the first is between two players whose movements are entered by the user, and the second mode is between a player whose movements are entered by the user and an automatic player who bases his movements on the MiniMax algorithm, While the third situation is the exact opposite of the second situation. The winning condition in this game is when any of the players places 5 consecutive blocks without any intersection, either horizontal, vertical or diagonal. When the game board is full of blocks without a player winning, the game ends.